Make on dashboard one single component
