+++ @3rdweb-sdk/react/hooks/useRegistry.ts
@@ -1,40 +1,29 @@
-import { contractKeys, networkKeys } from "../cache-keys";
+import { contractKeys } from "../cache-keys";
 import { useMutationWithInvalidate } from "./query/useQueryWithNetwork";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { useAddress, useSDK, useSigner } from "@thirdweb-dev/react";
-import { ChainId, SUPPORTED_CHAIN_ID, ThirdwebSDK } from "@thirdweb-dev/sdk";
-import { alchemyUrlMap } from "components/app-layouts/providers";
+import { useAddress, useSDK } from "@thirdweb-dev/react";
 import invariant from "tiny-invariant";
 
 type RemoveContractParams = {
   contractAddress: string;
-  contractType: string;
-  chainId: ChainId;
 };
+
 export function useRemoveContractMutation() {
-  const signer = useSigner();
+  const sdk = useSDK();
   const address = useAddress();
-  const queryClient = useQueryClient();
 
-  return useMutation(
+  return useMutationWithInvalidate(
     async (data: RemoveContractParams) => {
-      invariant(signer, "must have an active signer");
+      invariant(address, "cannot add a contract without an address");
+      invariant(sdk, "sdk not provided");
 
-      const { contractAddress, chainId } = data;
-      const rpcUrl = alchemyUrlMap[chainId as SUPPORTED_CHAIN_ID];
-      const sdk = ThirdwebSDK.fromSigner(signer, rpcUrl);
+      const { contractAddress } = data;
 
       const registry = await sdk?.deployer.getRegistry();
-      const tx = await registry.removeContract(contractAddress);
-      return tx;
+      return await registry.removeContract(contractAddress);
     },
     {
-      onSuccess: (_data, _variables) => {
-        const { chainId } = _variables;
-        return queryClient.invalidateQueries([
-          ...networkKeys.chain(chainId as SUPPORTED_CHAIN_ID),
-          ...contractKeys.list(address),
-        ]);
+      onSuccess: (_data, _variables, _options, invalidate) => {
+        return invalidate([contractKeys.list(address)]);
       },
     },
   );
